refactor(recs): moving recs to `views`
