+++ src/components/views/doorhanger/recommendations/recommendations.js
