--- src/containers/save/recommendations/recommendations.scss
@@ -1,21 +1,0 @@
-@import 'common/_styles/colors';
-@import 'common/_styles/variables';
-
-.recommendations {
-  background-color: $powder;
-  border-color: $white;
-  border-radius: 4px;
-  border-style: solid;
-  border-width: 2px;
-  box-shadow: $panel-shadow;
-  box-sizing: border-box;
-  color: $pitch;
-  font-family: $fontstack-default;
-  font-size: 13px;
-  font-weight: 300;
-  margin-top: 10px;
-  padding: 5px 0 0;
-  position: relative;
-  text-shadow: none !important;
-  width: 320px;
-}
