+++ src/containers/save/save.container.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react'
 import Toolbar from './toolbar/toolbar.main'
-import Recommendations from './recommendations/recommendations.main'
+import Recommendations from 'views/doorhanger/recommendations/recommendations'
 import Survey from './survey/survey.main'
 import { TransitionMotion, spring, presets } from 'react-motion' //
 import styled from '@emotion/styled'
